requirements.txt
psycopg2-binary
django
djangorestframework
pytest
pytest-django
coverage
black
ruff
whitenoise
colorlog
